Add files via upload
This was my first python project made when i started learning python. This is a simple quiz game that allows the users to answer based on the questions asked. In the end, the game evaluates the answers and gives feedback on each answer and finally calculates and displays the total score in the end. 
There was use of conditional statements (IF-ELSE) to evaluate the response of the user and determine if the answer is right or wrong. 
Implementation of case-insensitive comparisons and handling multiple correct answer string formats. The code also takes the user input providing a real time feedback which creates an engaging experience.