Add files via upload
MADLIBS is a simple and entertaining python project game that creates a fun story by taking the user for specific words and inserting them into a predefined template. The project demonstrates user interaction and basic string manipulation.
There is a utilization of basic syntax and input/output operations and string handling. Then there is good interaction with the users through user's responses in a meaningful plot. There was also a good use of string concatenation that joins with predefined texts. This project was really useful for me as a beginner when I coded this back then, it helped me grasp the basic concepts and proceed further with the learning.