Add files via upload
This project is called the guessing game where the user who is playing it tries to guess a randomly generated number between 1 to 100. The program provides hints on each guess to help the player get the correct answer.
This was the first project where I have used the random module (random.randint()) to generate a number within the range provided. There was a utilization of while loops as well to prompt the user until the correct guess is made for the loop to get exited. I have also  used conditional statements(if-elif) to compare the guessed number to the actual generated number and provide appropriate feedback.