Add files via upload
The random project generator is a python based project that generates random passwords based on user-defined criteria. The user can specify the number of letters, symbols and numbers they want in their password, and the program will generate the password accordingly. 
There was a use of lists, loops and manipulation of strings to generate the password.