Add files via upload
This project is a simple user interactive Rock-Paper-Scissors game implemented in Python. The user plays against the computer, which randomly selects its moves and determines the win in the end. 
There was usage of while loop to enable continuous play and if else condition to handle the logic of the game. There was of usage of random.randint() to randomly select the computer's move. The string manipulation '.lower' is used to handle case variations when the user is inputting the information.