Add files via upload
The dice roll project application that is made rolls the dice and the users can decide the number of dice to roll and continue until they decide to stop.
The random module was used again to give a number from  1-6 on the Dice. The conditional logic is implemented to provide different outcomes based on user input. Mainly there was use of os.system to clear the terminal screen for a better user experience. The while loop also allow the dice to roll repeatedly and enage continuously,