Add files via upload
This is a console induced tic tac toe game that is made using the python language. In this game there are two users who play against each other on a 3x3 grid aiming to align X OR O in a row, column, or diagonal.
The game board updates after each move and announces the winner when the win criteria is satisfied. The players input their moves on the grid in their selected positions by choosing a number between (0-8) given on each grid which replaces the number with an X or an O.