Add files via upload
The project is a simple countdown timer. It is a user input based timer where the user can provide the time from which it begins to run. The input time in seconds is converted to a minutes:seconds format using the divmod. The while loop decrements the time by one second after each iteration and the current time displayed in the format of MM:SS.
The sleep function which is time.sleep(1) pauses the loop for one second between each decrement to reproduce the passage of the real time. The "lift off" is printed to indicate the end of the timer. This project helped me learn more about time manipulation while also mastering my loops and input/outputs.