Add files via upload
This is an email validator project where the user puts their input and the program checks whether the given prompt is a valid email or invalid. The program first checks the length of the mail to make sure it is atleast 6 characters long. Then it checks if the first character is alphabetic and if the email contains exactly one "@" symbol. Then the dot placement check ensures that the email follows the correct conventions while naming a domain. It also loops through each character of the email to check for invalid characters or issues that occur during formatting.  If found any, the user is immediately notified about it and asked to correct the mail and try again. This happens until the user enters the right mail.