Add files via upload
This is an advancement of the python project where I started learning the module called openCV also known as computer vision while using the pre-trained haar cascade classifier to detect faces from a webcam feed in real-time.
The code first loads a pre-trained Haar Cascade Classifier using cv2.CascadeClassifier('haarcascade_frontalface_default.xml'). This file contains data that allows the classifier to recognize faces. Then the cv2.VideoCapture(0) method is used to access the webcam and 0 refers to the default camera device. To capture the frames, webcam.read() function  captures a frame from the webcam in a loop. Each frame is then converted to grayscale using the cv2.cvtColor(). The grayscale conversion is usuallt done to simplify and make it easier to the face detection process. The "face_cascade.detectMultiScale()" method detects the faces in the image. The grayscale image is taken in 2 parameters, firstly being the parameter 1.5 which is a scale factor that compensates for the size of the face at different distances from the camera. The second parameter which is 4 is the minimum number of neighbours that the detected rectangle shape should have to be considered a valid face. The method usually returns a list coordinates for the detected faces that is in the format: (x,y,w,h) which shows the top-left corner, width and height of the face bounding the box. For each detected face, the code draws a green rectangle around it using the function cv2.rectangle() function. Then the image is displayed in a window which is named as "face detection" using the "cv2.imshow()". You can simply exit the loop by pressing the esc key which is also represented with number 27 in ascii values. Once you exit the loop, the webcam is released using the function: webcam.release() to free the resources. Then finally, cv2.destroyAllWindows() is called to close the opencv windows.