Add files via upload
The code below is used to detect and decode the QR and barcodes in real-time using webcam. It is a very interesting project as any code displayed on the screen and is detected, its content is shown later and highlighted on the terminal.
The library modules used here are opencv and numpy and another specific library called pyzbar that is only used for decoding qr and bar codes. The code simply begins with an opening of the webcam using cv2.videocapture(0) which accesses the default webcam and starts to capture each frame of the video stream in real-time. Then the frame is captured inside a "while True" loop from the webcam using web_cam.read(). Success is a boolean indicating if the frame was captured correctly and the 'frame" is the actual image from the webcam. The decode(frame) function from the pyzbar library is used to detect QRcodes and barcodes within the current frame. So, for the each code detected, the "code.data.decode('utf-8') decodes the data in the QR or the bar code and prints it on the console. The coordinates of the BC or QR are used to draw polygon around the detected code with np.array([code.polygon], np.int32). The cv2.polyline draws the shape around the code. While displaying the result, the processed frame is displayed in a window using the cv2.imshow('Image frame', frame), showing the real time video feed with the detected code.
Then the loop continutes to run until the user presses the exit key and here it is the character 'q' , which breaks the loop and ends the program. Then finally the regular clean up is done with the cv2.destroyAllWindows().