Add files via upload
This code captures video from the webcam in real time and detects faces using the Haar Cascade Classifier which is a pre installed program. It then applies a Gaussian blur to any detected face, effectively anonymizing it. If no faces are detected, a message is displayed indicating an error. This program allows the user to exit the loop by pressing the 'e' key. 
The program begins by opening the webcam using cv2.VideoCapture(0). This catches the video from the default webcam, and camera.read() fetches each frame of the video. The Haar Cascade is a machine learning object detection method used to detect objects in images, in this case, faces and is loaded using the "cv2.CascadeClassifier('haarcascade_frontalface_default.xml')".
Inside the loop, faceCascade.detectMultiScale(frame, 1.2, 4) detects faces in the current video frame and the parameters 1.2 and 4 represent the scaling factor and  minimum number of neighbors required for the formation of the rectangle around a face respectively. The function also returns a list of rectangles(x,y,w,h) corresponding to the detected faces' coordinates and dimenions namely height and width. The ROI (face) is then blurred using cv2.GaussianBlur(ROI, (91, 91), 0). The (91, 91) specifies the size of the kernel , which controls the intensity of the blur for instance, larger values produce a stronger blur.
The blurred face is then placed back into the frame at the same position using frame[y:y+h, x:x+w] = blur_face. If no faces are detected in the frame then there is  a message 'error detecting face' displayed on the video feed using the cv2.puttext(). The text has a font called hershey complex small. Then the modified frame, which contains the blurred faces or error message, is displayed using the cv2.imshow('Face Blur', frame). Then it is followed by exiting the loop and cleaning up the windows using cv2.destroyallwindows().