Add files via upload
This is JOI, a speech-speech AI model that is made with very limited prompts which can be added later on. It is a simple AI chatbot that enables the user to talk to the bot or ask certain questions or ask to perform certain tasks. If the source asked for is in reach, it will definitely give the reply. For now this is sampled and made it easy to use with limited vocabulary and prompts. 
This uses speech_recognition library to recognise the commands through the mic. The pyttsx3 is used to convert the text back to speech, allowing the model to talk back to the user. The pywhtkit is used to perform various tasks, such as playing songs on youtube which was demonstrated on this code. I have also used wikipedia to fetch information especially when the user asks questions that starts with "who is". It can also fetch jokes using the library pyjokes when asked for one. 
The talk (text) function takes the text in the form of string and converts it into speech using pyttsx3 library and plays it back. It accesses the second voice available from the system's text-to-speech engine by setting the voices[1].id. The take_command() function captures the speech and converts it into text using the google speech recognition api and gives back the output in lowercases. The run_joi() function processes the voice commands from the user by calling the take_command() function. It checks if specific phrases (eg, play, time, who is etc) are present in the command and performs corresponding actions such as play songs, tells the time, provides insights using wikipedia, jokes or responds to personal queries. The assistant runs on an infinite loop which enables continuous listening for commands and responding until the program is manually stopped. This is my first basic chatbot project I've worked on and was a great experience learning about new libraries and modules.